Use nengo core SimPES operator
If it is available, rather than nengo-dl's version.

Set minimum Nengo version to 2.8.0

Move nengo backwards compatibility logic to compat.py
