+++ .appveyor.yml
@@ -11,7 +11,7 @@ environment:
     - MODE: test # default build
     - NENGO_VERSION: git+https://github.com/nengo/nengo.git
       TF_VERSION: tensorflow --pre
-    - NENGO_VERSION: nengo==2.7.0
+    - NENGO_VERSION: nengo==2.8.0
       TF_VERSION: tensorflow==1.4.0
       NUMPY_VERSION: numpy==1.14.5
       PYTHON_VERSION: C:\Miniconda35-x64
