+++ .nengobones.yml
@@ -127,7 +127,7 @@ travis_yml:
       python: 3.7
     - script: test-coverage
       env:
-        NENGO_VERSION: nengo==2.7.0
+        NENGO_VERSION: nengo==2.8.0
         TF_VERSION: tensorflow==1.4.0
         NUMPY_VERSION: numpy==1.14.5
       python: 3.5
