+++ .travis.yml
@@ -52,7 +52,7 @@ jobs:
     python: 3.7
   -
     env:
-      NENGO_VERSION="nengo==2.7.0"
+      NENGO_VERSION="nengo==2.8.0"
       TF_VERSION="tensorflow==1.4.0"
       NUMPY_VERSION="numpy==1.14.5"
       SCRIPT="test-coverage"
