+++ nengo_dl/compat.py
@@ -1,7 +1,13 @@
-"""Utilities to ease cross-compatibility between TF 1.0 and TF 2.0."""
+# pylint: disable=unused-import,ungrouped-imports
+
+"""
+Utilities to ease cross-compatibility between different versions of upstream
+dependencies.
+"""
 
 from distutils.version import LooseVersion
 
+import nengo
 import tensorflow as tf
 
 if LooseVersion(tf.__version__) < LooseVersion("2.0.0"):
@@ -42,3 +48,92 @@ else:
         """Always returns RefVariables instead of (maybe) ResourceVariables."""
 
         return tf.compat.v1.Variable(*args, use_resource=False, **kwargs)
+
+
+if LooseVersion(nengo.__version__) < "3.0.0":
+    class SimPES(nengo.builder.Operator):  # pylint: disable=abstract-method
+        r"""
+        Calculate connection weight change according to the PES rule.
+
+        Implements the PES learning rule of the form
+
+        .. math:: \Delta \omega_{ij} = \frac{\kappa}{n} e_j a_i
+
+        where
+
+        * :math:`\kappa` is a scalar learning rate,
+        * :math:`n` is the number of presynaptic neurons
+        * :math:`e_j` is the error for the jth output dimension, and
+        * :math:`a_i` is the activity of a presynaptic neuron.
+
+        Parameters
+        ----------
+        pre_filtered : Signal
+            The presynaptic activity, :math:`a_i`.
+        error : Signal
+            The error signal, :math:`e_j`.
+        delta : Signal
+            The synaptic weight change to be applied,
+            :math:`\Delta \omega_{ij}`.
+        learning_rate : float
+            The scalar learning rate, :math:`\kappa`.
+        tag : str (Default: None)
+            A label associated with the operator, for debugging purposes.
+
+        Attributes
+        ----------
+        pre_filtered : Signal
+            The presynaptic activity, :math:`a_i`.
+        error : Signal
+            The error signal, :math:`e_j`.
+        delta : Signal
+            The synaptic weight change to be applied,
+            :math:`\Delta \omega_{ij}`.
+        learning_rate : float
+            The scalar learning rate, :math:`\kappa`.
+        tag : str (Default: None)
+            A label associated with the operator, for debugging purposes.
+
+        Notes
+        -----
+        1. sets ``[delta]``
+        2. incs ``[]``
+        3. reads ``[pre_filtered, error]``
+        4. updates ``[]``
+        """
+
+        def __init__(self, pre_filtered, error, delta, learning_rate,
+                     encoders=None, tag=None):
+            super(SimPES, self).__init__(tag=tag)
+
+            self.pre_filtered = pre_filtered
+            self.error = error
+            self.delta = delta
+            self.learning_rate = learning_rate
+            self.encoders = encoders
+
+            # encoders not used in NengoDL (they'll be applied outside the op)
+            assert encoders is None
+
+            # note: in 3.0.0 the PES op changed from a set to an update
+            self.sets = [delta]
+            self.incs = []
+            self.reads = [pre_filtered, error]
+            self.updates = []
+
+        def _descstr(self):
+            return 'pre=%s, error=%s -> %s' % (
+                self.pre_filtered, self.error, self.delta)
+
+    # remove 'correction' from probeable attributes
+    nengo.PES.probeable = ("error", "activities", "delta")
+
+    class Convolution:
+        """Dummy Convolution class."""
+
+    class ConvInc:
+        """Dummy ConvInc class."""
+else:
+    from nengo.builder.learning_rules import SimPES
+    from nengo.builder.transforms import ConvInc
+    from nengo.transforms import Convolution
